Custom Armor Framework 99%
Near fully written out custom armor framework. There's something wrong in the AAX section but the project is nearly at the finish line. The problem: For both the custom CH252 helmet and the custom uniform for marines, only 1 specific arid variant is available in ace arsenal and aax is not hooking in.